Add the ability to optionally list parameters for CMZ protocols
You can now write:

CMZProtocol! { proto_name<param1, param2>,
  [the rest is the same as before]
}

to say that param1 and param2 are public constants (Scalars) that will
be specified at runtime.  Then param1 and param2 can appear in the
statements in the statement list in the protocol declaration.

The "<param1, param2>" is optional, and defaults to specifying no
parameters.