Use sigma_compiler_prover or sigma_compiler_verifier as appropriate
If the Cli version of the macro is invoked (only output the client
code), use sigma_compiler_prover for the client proof and
sigma_compiler_verifier for the issuer proof.

If the Iss version of the macro is invoked (only output the issuer
code), use sigma_compiler_verifier for the client proof and
sigma_compiler_prover for the issuer proof.