Change nomenclature of "params" to "instance" for the underlying sigma protocol
The term "params" is now used only for the public parameters to the CMZ
protocol definition (for example, a fee in the wallet example, which is
a public value chosen at runtime).

The set of public values in a sigma protocol is now termed an
"instance", to match the usage by the underlying sigma_rs crate.